requirements.txt
pillow==7.2.0
tensorflow~=2.6.0
keras~=2.6.0
matplotlib==3.3.1
mlflow~=1.27.0
